Avoid ugly padding next to scroll bar in Plasmoid
It is actually just the normal marging/padding Plasmoids normally have but
with the scroll bar it doesn't look good.

The margins would be automatically collapsed if the scroll view was the
immediate content item. However, that's not the case so we need to set the
according property manually using `Plasma.Representation`.