Fix checks for Plasma 6
The alpha defines the macros

```
#define PLASMA_VERSION_MAJOR 5
#define PLASMA_VERSION_MINOR 27
#define PLASMA_VERSION_PATCH 80
```

so the current version checks (which were created based on the macros from
a development build) are not working. I have no clue how these numbers are
crafted so it is supposedly the easiest to just check for Qt 6 as Qt itself
uses a sane versioning system.