Make it more clear when the options regarding metered connections are not available
Since this feature will likely not be available in mingw-w64-based builds of
the next release (due to missing cppwinrt) it is important to make it clear
when those options don't actually work. With this change the checkboxes are
completely disabled in case the feature is not available (instead of just
providing the information via the tooltip).