Fix re-connect when process status is considered and Syncthing restarted
When Syncthing restarts itself, its main process is not terminating (as its
main process is just a monitoring process). The "Consider process status …"
feature was relying on the process being stopped and started again as
trigger for trying to re-connect, though. To make it also work when
Syncthing restarts itself, this change checks the log for an exiting
message of the monitoring process and considers the process running again
when the GUI address is logged again.

This fixes https://github.com/Martchus/syncthingtray/issues/236.

Notifications are not (yet) suppressed. (Not sure how easy it is and how
much sense it makes.)