Postpone connecting until pending requests have finished
Otherwise a handler for a pending connection from before the call to
`connect()` might be handled after the call to `connect()` and set the
connection back to the disconnected state. This is already done in
`reconnect()`. If `connect()` is postponed and then `reconnect()` is
invoked as well then `reconnect()` will win.