Update overall "Out of Sync" status when a dir changes its status
The overall "Out of Sync" status relies on
`SyncthingConnection::hasOutOfSyncDirs()` and not just on
`SyncthingConnection::status()`. Therefore the `statusChanged()` event does
not cover all cases when the overall status needs to update. This usually
leads to a stuck "Out of Sync" status (until the status changes otherwise).

This change introduces the additional `hasOutOfSyncDirsChanged()` event to
cover these status changes. It is only emitted on demand (when
`hasOutOfSyncDirs()` has been called before) and when `statusChanged()` is
not invoked anyway to avoid updating UI compoents multiple times.