Clarify status of the Debian package in the Neon distribution.
Neon is downstream from Ubuntu, which is downstream from Debian.  In
Ubuntu and Neon most Debian packages are distributed in "universe"
repository.  Neon replaces Debian-provided Qt and KDE/Plasma with
their own packages, because the focus of Neon is to be the
distribution with the most up-to-date Qt and KDE/Plasma.

Unfortunately this breaks a subset of the software in "universe";
Syncthing Tray is one package that is broken by this policy.  The
subset of software in "universe" broken by this policy can be fixed by
rebuilding it against the Neon-provided copy of Qt and KDE/Plasma;
however, configuration changes or patching may sometimes be required.
Neon considers resolution of this situation to be out-of-scope for
their distribution.

See #254
and https://bugs.kde.org/show_bug.cgi?id=491044