Workaround context menus not showing anymore as of Qt 6.8
At least the upstream Windows builds of Qt 6.8.0-beta2 and Arch Linux
builds of Qt 6.8.0-beta3 do not trigger `customContextMenuRequested()`
anymore for widgets that are nested within a popup. I noticed this when
testing on Windows 11 and on KDE Wayland. It does not work to use
`Qt::DefaultContextMenu` and `QWidget::contextMenuEvent()` as well.

This problem occurs with different styles and it is probably not platform-
specific.

The problem goes away when showing the tray menu as a regular window (and
not as a popup). So it is probably related to this specific usage of menus
in Syncthing Tray.

This change works around the issue by showing the context menu via the
mouse release event handler. (Supposedly it should be the mouse press
event on certain platforms like X11/Wayland so this probably still needs
further tweaking.) On the upside this means the positioning hack can be
removed from the `showViewMenu()` function. I tested this with Qt 6.8.0
beta3 (no hack was required anyway) and Qt 5.15.14 (hack was required).