Increase default limit for disk event query to 200 and let model follow
The model already keeps at most 200 disk events around which it not an
unreasonably high number. To simplify the code and to improve consistency
it makes sense to just use the query limit in the model and increase that
to 200. The query will hopefully not be too expensive and now it is
completely avoided anyway unless somebody actually looks at the recent
changes.