Fix handling row count in file model
Before this change, the model returns 1 as row count for directory items
that have not been populated yet. This made sense because this way the tree
view allows expanding the item and possibly fetches more data. However, it
is technically wrong as no valid index exists. This is not a problem for
the tree view from Qt Widgets but the tree view from Qt QuickControls does
not handle it well and shows surplus empty rows due to it.

This change simply inserts a dummy item. This is done before the query
anyway for the sake of showing a loading indication but now `rowCount()`
already makes sure such an item exists.