Fix race condition in models test
The test only worked if the status was processed before the config. With
this change it should not matter anymore which one is faster. Since
applying the raw config is problematic to do immediately after reading the
config (because the status might still be missing) this change removes this
call; it needs to be done manually anyway to avoid the race condition.