Ensure the models test is not disturbed by local files
Apparently the three dots are resolved to the current working directory
under Windows leading to test failures (as many local files from the build
directory show up in the model due to this). With this change this is
avoided. It would actually make sense to test the local lookup and set this
to an existing but well-defined directory.