Fix waiting for connection when applying settings in wizard
This broke when making `setAutoReconnectInterval()` emit the
`statusChanged()` event because the resulting disconnected event will be
consumed by wizard-related code leading to "Unable to connect …" showing up
on the wizard. This event was added so the tooltip will be updated when the
interval changes. This commit adds an explicit event for that.