Allow keeping track of sleep/wakeup time in `SyncthingProcess`
So it has the same API as `SyncthingService`. This way we no longer need to
distinguish the classes when handling settings and could implement
https://github.com/Martchus/syncthingtray/issues/27 via `SyncthingProcess`.

This commit moves all the actual implementations of the `isActiveFor()` and
`isActiveWithoutSleepFor()` to a single place (`SyncthingProcess`) making
the other functions in `SyncthingService` and `SyncthingLauncher` mere
wrappers.