Improve column sizing in QTreeView
This especially helps in case folder IDs/labels or device names are very
long. In this case the first column uses almost all of the space. That is
problematic for nested items where the second column should normally be the
bigger one.

This changes fixes the problem by simply using only one column and
rendering all the text in the custom item delegates.

At this point no text eliding is done; this should probably be improved in
the future.