Try to improve opening path on Android
This still doesn't seem to work with Cx File Explorer but at least the code
now shows a toast for easier debugging and uses the canonical path which
might be less problematic. There's now also a comment explaining why
Uri.fromFile doesn't work. (Not sure why it works for the Syncthing-Fork
app.)