Hide "Downloads" tab by default
The "Downloads" tab is most of the time empty despite downloads going on,
see https://github.com/Martchus/syncthingtray/issues/293.

It actually needs quite some rework to be useful but most users probably
don't use it anyway. Not even the official GUI shows the download progress.
So for now let's hide it by default.

It would make sense to extend the GUI to be able to show out of sync items
like the official GUI does. Maybe this could replace the "Downloads" tab in
the future.