Handle destruction of activity more gracefully
Since Qt will put the app in a state where it can't recover from (the UI
can never be shown again) we have to terminate everything else as well.
Otherwise Go threads and the service/notification would still be around but
one could never access the UI again.

This is of course not ideal. Ideally, the activity lifecycle should be
independent from the service livecycle and the service should be able to
outlive the activity. The activity should be able to re-create itself once
destroyed.