Use default system font under Android
With this change the Android app uses the font family configured in the
system settings under Android (and not always Roboto).

This is still not ideal as font files for different weights need to be
loaded manually and at least on an Samsung device those all seem to
resolve to the same path (for instance
`/data/app_fonts/0/Foundation/DroidSans.ttf`). Even when loading those
fonts from data Qt doesn't load the files for extra font weights.

When the standard font is selected the path is
`/system/fonts/Roboto-Regular.ttf` and all font weights are available.