Replace "Initializing …" with more useful notification text
Now that launching the activity from the background works it is possible to
start Syncthing from the notification in that state.

Of course it would be better to simply initialize the app while in the
background (without starting the activity) but this is not possible without
further modifications in Qt.