Remove hack to ensure font functions are available
The hack leads to linking against `libjnigraphics.so` unnecassarily and
also including and invoking the bitmap conversion code unnecassarily.

So this change replaces it with another hack that resolves the symbols
dynamically via `dlopen()` and `dlsym()`. (Just loading the library via
`dlopen()` or `System.loadLibrary()` doesn't work, also when `RTLD_NOW` is
specified.)