Do not wrongly clear the "manually stopped flag"
Follow-up of 0ba4dfca to make it work not just sometimes but hopefully
consistently. The problem is that `handleLauncherStatusBroadcast()` can be
called between flagging the manual stop and the actual termination if just
one of the other launcher status values change. So the flag should only be
clear when actually transitioning from not running to running.