Fix running into assertion error when terminating after web UI was shown
The destructor of the QObject-based TrayWidget will delete all its
children including the web view dialog. Before that it is already
"destructed enough" to no longer count as TrayWidget for the sake of
invoking the `handleWebViewDeleted` slot leading to a failing assertion.
This change therefore removes the connection of that slot.

This is *not* required for SyncthingApplet because it is not the parent of
its web view dialog (and deletes its dialog manually).