Improve Edge-to-edge support in Qt Quick GUI for Android
Now everything works with a recent enough Qt. The only not so nice part
that remains is the coloring of icons in the status bar when light mode is
used. These icons needed to be dark when over the drawer or bright when
over the menu. Maybe the Qt Quick GUI shouldn't use a dark menu even in
light mode.