Improve German translation for "No remote device connected"
The correct wording doesn't matter much. However, "Nicht mit anderen
Geräten verbunden" is already used for "Not connected to other devices" and
it looks bad that both are the same.

Maybe "Not connected to other devices" should be completely avoided if
the overall status is "No remote device connected" as it is quite redundant
even if the wording is not identical.