Fix showing needed items in Qt Quick GUI
With a QML debugger session the needed page is working without this change.
Hence I haven't discovered this problem before. It probably was working
also without a QML debugger session at some point with an older version of
Qt. However, with Qt 6.10 `params` is always `undefined` (under GNU/Linux
and Android). This change fixes the problem by simply moving the code to
compute the params into the function.