Avoid showing not so meaningful initial download speed when opening UI
When the UI is hidden the polling for traffic statistics is suspended. When
then opening the UI and polling is continued, so far a download speed based
on the last statistics from before the polling was suspended was computed.
This download speed isn't telling much about the download speed right now
and thus rather confusing. This change prevents this by computing the speed
only while polling.