Ensure Android service notification is shown again after dismissed
It is possible to dismiss the notification which is ok. However, it should
be possible to show the notification again. With this change the
notification is shown again when the app is opened again. Maybe a "Show
notification again" action somewhere in the app would make more sense but
for now this is good enough.